🎨 Clean up main navbar component
